requirements.txt
django
Pillow
django_rest_framework
requests
psycopg2
djangorestframework
drf-yasg
channels
channels_redis